+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-05-01
+### Changed
+- [#8263: Flaskアプリに変更して、ジョブ登録はFlaskで定義したWebAPIで受け付ける](https://redmine.u6k.me/issues/8263)
+- [#8334: 当日のレース情報をクロールして、再クロール＆投票処理を予約する](https://redmine.u6k.me/issues/8334)
+
 ## [1.1.0] - 2020-04-18
 ### Fixed
 - [#8249: カレンダー上の当日レースは、RaceRefundではなくOneDayRaceListにリンクされるため、当日レースが取得できていない](https://redmine.u6k.me/issues/8249)